requirements.txt
# RAG Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.3.1
qdrant-client==1.7.0
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
mangum==0.17.0

# Utilities
httpx==0.25.2
